Move the parser package into the compiler package.
