+++ .gitattributes
@@ -1,7 +1,7 @@
-internal/vm/parser/parser.go -diff -merge
-internal/vm/parser/parser.go linguist-generated=true
-internal/vm/parser/y.output -diff -merge
-internal/vm/parser/y.output linguist-generated=true
+internal/vm/compiler/parser/parser.go -diff -merge
+internal/vm/compiler/parser/parser.go linguist-generated=true
+internal/vm/compiler/parser/y.output -diff -merge
+internal/vm/compiler/parser/y.output linguist-generated=true
 logo.png binary
 internal/mtail/logo.ico binary
 .gitattributes export-ignore
