+++ cmd/mdot/main.go
@@ -31,7 +31,7 @@ import (
 	"github.com/google/mtail/internal/mtail"
 	"github.com/google/mtail/internal/vm/compiler/ast"
 	"github.com/google/mtail/internal/vm/compiler/checker"
-	"github.com/google/mtail/internal/vm/parser"
+	"github.com/google/mtail/internal/vm/compiler/parser"
 )
 
 var (
