+++ internal/vm/compiler/checker/checker_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/google/mtail/internal/testutil"
 	"github.com/google/mtail/internal/vm/compiler/ast"
 	"github.com/google/mtail/internal/vm/compiler/checker"
-	"github.com/google/mtail/internal/vm/parser"
+	"github.com/google/mtail/internal/vm/compiler/parser"
 	"github.com/google/mtail/internal/vm/symbol"
 	"github.com/google/mtail/internal/vm/types"
 )
