+++ internal/vm/compiler/parser/driver.go
