+++ internal/vm/compiler/parser/lexer.go
