+++ internal/vm/compiler/parser/lexer_test.go
