+++ internal/vm/compiler/parser/parser_test.go
