+++ internal/vm/compiler/parser/tokens_test.go
